Reduced consistency checking to minimum.
This turns out to be a much-needed optimization right now for some reason - stuff's lagging all over the place, and this stops it.

As it is, the only thing to be worried about is stuff that actually pokes at it's subelements, which now have to declare as such to get the proper notifications.