Changes all over the place. Image system in gabien changed tons.
Here's the specifics.
gabien-common: Images no longer act DirectDraw-like (separate magical colour-key controlled by blit function), instead the colour-key is controlled by how you load the image (as it really was all along).
gabien-javase: Implement this.
This makes it a LOT easier to bring in the changes to:
gabien-app-r48: Firstly, temporarily disable MoveCommand serialization since it doesn't work on the R2k Test Game. I'm going to guess it was useful of me to write a structure dumper, since I can output the blob as base64.
Secondly, PNG8IImageLoader extracts the palette to get the right colour key. This is because I tried to actually read the IDAT, and failed miserably. This is much *less* brittle, I believe.
And of course, to make that work, I had to switch over the image rendering API so that the tile/event renderers wouldn't have to know if the image they received had alpha as CK or as normal alpha.
My hope is, this will support both 8-bit indexed PNGs (by reading the palette) and the EasyRPG 'standard' 32-bit ARGB PNGs (by simply using the standard ImageIO stuff) without issue.

...also, this is going to do a number on my already completely broken version of the gabien runtime for Android, but that's fine because I had nothing good that would work on mobile anyway.