Clean up some little messes.
Ok, so, NOW I can finish the command list and get to the LDB.
Right? Right?
For future reference, that Force Push where I changed history but made it quite obvious *where*? That was me quietly cleaning up BM8I. Technically it was just a palette, but I decided, "fuck it, treating the palette as an independent resource now." At this point the only thing that remains is a vanilla BMP header. Not even history of it.
As it is, IkaObjectBackend's serialization helpers were nabbed from my older IkachanMapEdit system. This ought to explain a lot about it to you, like why they're even in separate classes when R48ObjectBackend is bigger and in one.