Finish up changes to tileset modification
This in theory allows the same tileset flags specification system to apply to cases where the tileset is resizable, while denying resize in cases where it shouldn't be resizable.
It also provides 'manual editing' (the plain UINumberBox) in case I missed anything in the bitfields and thus someone needs to enter numbers manually.