Fix dictionaries not updating properly, r2k Control Switches & Conditional Branch name issues, and make the panorama implementation emulate r2k.
The specifics of the panorama implementation details are:
 1. Tile if anything is looping (since when nothing is looping it's *usually* a parallax-as-map. This could probably be considered a heuristic.)
 2. If anything is looping, get the centrepoint of a 20x15 tile 'camera' (the parallax basis), get the difference between that and the actual camera's centrepoint, and then calculate parallax from that.
 Assuming the parallax 'basis' in RM2k is indeed pinned to the top-left (as seems to be the case),
  this appears to precisely emulate the behavior.

As it is, I am kind of worried that there *may* be a 2k3 data corruption bug in R48. Or maybe there isn't. Or maybe there was, and now there isn't.
If there is, it's in the LDB file and it affects the igm commands.