Fix data corruption with @menu_commands_2k3 on 2k targets in 2k3 RPG_RT
Ok, so, what goes on here is:
The 2k3 RPG_RT assumes it's always running a 2k3 game.
If the @menu_commands_2k3 member is missing, you have one blank space, then the End Game option.
If it exists, you just have the End Game option.
If it exists and has a blank, you somehow have (and I keep wondering how this is possible) *two* blanks and the End Game option.