Get rid of the static variable 'nextMapTool'
nextMapTool is what forces R48 to only edit one map at a time.
Getting rid of it sounds good to me, even if I can't think of a way to make the UI work.
My current idea is to have a button to open the map in a separate window, and clip that
 onto the toolbars via a UIAppendButton.