one last bit of build confusion
Information for anyone who receives build TAK-IMI-TEST and wonders why the IMI installer has a different SHA than it should: I retroactively applied the effects of this commit to it. And I might have forgotten to make it properly reproducable. Will try and stripzip it soon, but I have enough build issues as it is.