Get rid of PointerConnector, fix UITreeView bug, introduce text feedback on the R48 side
This is one of those cases where I can happily say I got rid of a few classes.
Though unfortunately more classes had to be added to deal with resulting problems.
I could probably save a bit of code by making the pointer receivers into an array, odd as it might be.