Fix IMI breakage due to DM2 implementation issues
IMI is probably going to get used as an object representation format,
 as it's easier to adapt to changes than .r48 files or binary data.

In particular I'm thinking of using it for R2k templates and the like.
A current problem with this is how to handle object initialization nicely.

Current idea:
Auto-instantiation of objects with known classes is a quick-fix but is a practical low-code solution for what is ultimately a convenience system.